Change theme to helper object instead of from themeprovider
